rename PrayingScreen to PrayerScreen and move it to the new /screens directory
