+++ packages/apollos-ui-prayer/src/screens/index.js
@@ -1,3 +1,4 @@
 export AddPrayerScreenConnected from './AddPrayerScreenConnected';
 export ConfirmationDialogScreen from './ConfirmationDialogScreen';
 export PrayerDialogScreen from './PrayerDialogScreen';
+export PrayerScreen from './PrayerScreen';
